config.sh file
Correccion

+++ config.sh
@@ -7,7 +7,7 @@ clean()
    for file in *
    do
       #echo $file
-      if [ -f $file -a -x $file -a $file != "config.sh" ]
+      if [ -f $file -a -x $file -a $file != "config.sh" -a "$file" == "${file/\./}" ]
       then
          echo "rm -f $file"
          rm -f $file
